requirements.txt
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1763585112330/work
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.0.1
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1762777100038/work
django-cloudinary-storage==0.3.0
django-jazzmin==3.0.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1734007102193/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
urllib3==2.5.0
whitenoise==6.11.0
